update, delete: 북마크 관련 라우트 핸들러 로직 작성, 필요 없는 라우트 핸들러 파일 삭제
